refactor: Update spinner design
