--- src/views/components/common/Spinner/Spinner.module.scss
@@ -1,23 +1,0 @@
-@use 'src/views/styles/colors.module.scss';
-
-$spinner-border-width: 10px;
-
-.spinner {
-    border: 1px solid colors.$charcoal;
-    border-width: $spinner-border-width;
-    border-top: $spinner-border-width solid colors.$tangerine;
-    margin: 0 auto 15px auto;
-    border-radius: 50%;
-    width: 60px;
-    height: 60px;
-    animation: spin 1s linear infinite;
-}
-
-@keyframes spin {
-    0% {
-        transform: rotate(0deg);
-    }
-    100% {
-        transform: rotate(360deg);
-    }
-}
